+++ src/pages/DarkHomePreview.tsx
@@ -122,7 +122,7 @@ const DarkHomePreview = () => {
               <div className="relative pb-[56.25%] h-0">
                 <iframe
                   className="absolute top-0 left-0 w-full h-full rounded-lg"
-                  src="https://www.youtube.com/embed/8WInnaFHMY0"
+                  src="https://www.youtube.com/embed/IFZdtPfUtPo"
                   title="Watch a Demo"
                   frameBorder="0"
                   allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"